pyproject.toml
[project]
name = "screendiary"
version = "0.1.0"
description = "Desktop activity tracker for Linux - periodic screenshots with OCR and AI search"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "numpy",
    "pytesseract",
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "python-multipart",
    "httpx",
    "openai",
    "click",
    "structlog",
    "pystray",
]

[project.scripts]
screendiary = "screendiary.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/screendiary"]
